Moves and build.gradle
